backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
qiskit>=1.0.0
qiskit-aer>=0.13.0
python-multipart
kagglehub
numpy>=1.24.0
scipy>=1.11.0
pydantic>=2.0.0
slowapi>=0.1.9
python-dotenv>=1.0.0
aiohttp>=3.8.0
qiskit-ibm-runtime>=0.20.0
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
aiomysql>=0.2.0
prometheus-client>=0.17.0
structlog>=23.1.0
httpx>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
dependency-injector>=4.41.0
APScheduler>=3.10.0
gunicorn>=21.2.0
